separated select form component from selectPackagePage
